requirements.txt
altgraph==0.17.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
CTkToolTip==0.8
customtkinter==5.2.2
darkdetect==0.8.0
et_xmlfile==2.0.0
msoffcrypto-tool==6.0.0
Nuitka==2.8.9
olefile==0.47
openpyxl==3.1.5
ordered-set==4.1.0
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.8
pep8==1.7.1
pillow==12.1.0
psutil==7.2.1
pycparser==2.23
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.10
pypdf==6.6.0
pypdfium2==5.3.0
pytesseract==0.3.13
Send2Trash==2.1.0
setuptools==80.9.0
watchdog==6.0.0
zstandard==0.25.0
